Rename the project to mingwGitDevEnv
That name is more appropriate as our git.exe is compiled using MinGW, not MSYS,
i.e. it does not depend on msys-1.0.dll. The MSYS environment is only required
by Bash etc. in order to run the scripts that come with Git.

This resolves also any possible confusion with msys-git, a project by the MinGW
community that strives towards creating a Git version compiled with MSYS that
e.g. works with mintty, see

http://permalink.gmane.org/gmane.comp.gnu.mingw.user/36756
http://permalink.gmane.org/gmane.comp.gnu.mingw.user/36766

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>