Add an mgwport file for mingw32-curl
The build method does not use "configure" because I have not found no way
to enable IPv6 and WINSSL with it.

SSL certificate validation is done using WINSSL and the Windows Certificate
Store, not with a shipped cacert.pem.

Build flags inspired by the recent changes to src/curl/release.sh in the
msysgit project, see

    https://github.com/msysgit/msysgit/commit/37e42ab860ae555d2ba4b0e3e014a0bc6860f0d5

Signed-off-by: Thomas Braun <thomas.braun@virtuell-zuhause.de>
Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>