installer: Fix parsing for groups that are not defined as a package child
This caused mingw32-unzip to erroneously be assigned to the "MinGW Project
Infrastructure" group of the "mingw32-wsl_rc" meta package which is not
part of any group hierarchy and thus was skipped when iterating over the
group hierarchy. Consequently, mingw32-unzip was never installed which
made any program use msys-unzip instead, which results in several Git
tests to fail.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>