Move the packages directory to a separate repository
This way it is easier for CI builds to detect changes in order to
automatically (re-)build packages. Moreover, the packages can only be
build from within a mingwGitDevEnv installation anyway, so it makes sense
to be able to directly commit changes to the packages from a
mingwGitDevEnv installation by cloning the repository there.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>