Store 32-bit color icons before 8-bit color icons in the file
Tk's "wm iconbitmap" seems to simply use the first 32 x 32 pixel icon it
finds in an icon file, so make sure we use the 32-bit color icon.

Note that this violates the guidelines at

http://msdn.microsoft.com/en-us/library/ms997636.aspx

which say "The icons should be ordered in increasing color resolution".

The icon order was changed with the "Greenfish Icon Editor Pro" Open
Source Software, see

http://greenfishsoftware.blogspot.hu/2012/07/greenfish-icon-editor-pro.html

For reference, we needed to do the same thing for msysgit, see

https://github.com/msysgit/msysgit/commit/fd6f8d03fb63dc1c160ff2dd7aeb62fd6682c742

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>