profile: Remove the current directory (".") from PATH
While this is the default in Windows, it is not on Unix, and MSYS is about
Unix after all. Moreover, this has been requested several times by users,
see e.g. [1]. Finally, having "." in PATH is particularly annoying if
there is a locally compiled "git.exe" in the root of the repository in the
"git" directory, as Git commands in that directory would then invoke
"/git/git.exe" instead of "/mingw/bin/git.exe".

[1] https://github.com/msysgit/msysgit/issues/248

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>