Install required and optional packages separately
This will allow to capture only required packages and their dependencies
in the future, which is what we need for the end-user installer.

Also use the term "optional" instead of "recommended" as the initial
selection of recommended packages can be changed, and thus does not
necessarily match the recommendation anymore.

Finally, remove the check for no packages being selected. This could have
never happened as we always have the list of required packages hard-coded.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>