Update project URLs as this project is now deprecated
The focus of Git for Windows is now the MSYS2-based environment. The
MSYS1-related repositories were moved out of the git-for-windows
organization on GitHub to not confuse contributors.

Also delete the "Getting in touch" section from the README in this context
as the mentioned resources should not be used for the MSYS1-based projects
anymore.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>